+++ src/components/AppCommandBar/SortButton.tsx
@@ -15,10 +15,10 @@ import {
   Icon,
   Text,
 } from "office-ui-fabric-react";
-import { SortField, SetSortParameters } from "../model/Order";
+import { SortField, SetSortParameters } from "../../model/Order";
 import { useDispatch } from "react-redux";
-import { setFieldsArray } from "../model/Subject/SetFieldsArray";
-import { setFieldsDesc } from "../model/Subject/SetFieldsDesc";
+import { setFieldsArray } from "../../model/Subject/SetFieldsArray";
+import { setFieldsDesc } from "../../model/Subject/SetFieldsDesc";
 import { BUTTON_HEIGHT } from "./Common";
 
 interface SortCalloutProps extends SetSortParameters {
