+++ src/views/components/ViewsNav/ViewsNav.tsx
@@ -12,7 +12,7 @@ import { useDispatch, useSelector } from "react-redux";
 
 import { RouteComponentProps } from "react-router";
 import { State } from "../../../Reducer";
-import { commandBarStyle } from "../../../AppCommandBar/AppCommandBar";
+import { commandBarStyle } from "../../../components/AppCommandBar";
 import { createView } from "../../../model/Views/CreateView";
 import { gotoView } from "../../../model/Views/Routing";
 import { subjectBase } from "../../../model/Subject/Routing";
